WIP: Split out runtime into a directory of files
